% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPEAR_wrapper_functions.R
\name{apear_data_prep}
\alias{apear_data_prep}
\title{Convert scDAPP Pathway Analysis Results to aPEAR Input Format}
\usage{
apear_data_prep(object, show_Category = TRUE, ...)
}
\arguments{
\item{object}{An object from scDAPP pathway analysis, expected to be an RDS
file located at \code{multisample_integration/pathwayanalysis_crosscondition/DE_pathways_plot_objects_list.rds}.
The object must be a list containing pathway analysis results structured
specifically for compatibility with this function.}

\item{show_Category}{Logical; if TRUE (default), includes the MSIGDB category in
the cluster description. This affects the transformation of pathway names based
on the presence of certain prefixes (e.g., "HALLMARK_").}

\item{...}{Additional arguments passed to the underlying clustering functions.}
}
\value{
Returns a nested list where each key represents a comparison condition
(e.g., \link{condition A}\emph{vs}\link{condition B}). Each corresponding value is a list of
data frames, one for each unique cell cluster identified in the analysis. Data
frames contain columns: "Description", "Category", "setSize", "enrichmentScore",
"NES", "pvalue", "p.adjust", "core_enrichment", and "Des_Full", detailing the
cluster's description, MSIGDB category, set size, enrichment score, normalized
enrichment score (NES), p-value, adjusted p-value, core enrichment genes, and
full description, respectively.
}
\description{
Converts gene set enrichment analysis (GSEA) results from scDAPP across
different conditions into a format compatible with aPEAR. It organizes data
from \code{pathway_analysis_mainlist_comp} by condition and cluster, arranging
MSIGDB categories (e.g., Hallmark) into a designated column. This function
requires \code{tidyverse} for data manipulation and string operations.
}
\details{
The function performs several key operations: validating the input object's
structure, creating a mapping of pathways to clusters, and transforming the pathway
analysis results into the aPEAR input format. Error handling is robust, with
specific messages provided for common issues such as missing or incorrectly
structured input data. It is crucial that the input object adheres to the expected
structure for successful transformation.
}
\note{
This function assumes that \code{tidyverse} is installed and available
for use. If not, please install these packages using \code{install.packages()}.
}
\examples{
# Assuming `pathway_analysis_result` is your scDAPP analysis result object:
apear_input_list <- apear_data_prep(pathway_analysis_result)

}
