% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPEAR_wrapper_functions.R
\name{apear_find_clusters}
\alias{apear_find_clusters}
\title{Select Pathways and Calculate aPEAR Clusters}
\usage{
apear_find_clusters(
  apear_input_list,
  initial_minCS = 4,
  max_label_threshold = 50,
  max_sign_pathway = 500,
  min_leading_edge_threshold = 2,
  output_dir = NULL,
  save_cluster_parameters_and_results = TRUE,
  ...
)
}
\arguments{
\item{apear_input_list}{The output list from the \code{apear_data_prep} function, structured by comparison groups.}

\item{initial_minCS}{The initial minimum cluster size for \code{aPEAR::findPathCluster}. Defaults to 4.}

\item{max_label_threshold}{The maximum number of labels allowed on the final graph. Defaults to 50.}

\item{max_sign_pathway}{The maximum number of pathways considered, based on p-value significance. Defaults to 500.}

\item{min_leading_edge_threshold}{The minimum number of leading edge genes required for a pathway. Defaults to 2.}

\item{output_dir}{The directory where the \code{findPathCluster} results will be saved. \code{NULL} by default.}

\item{save_cluster_parameters_and_results}{Logical; if \code{TRUE}, the cluster parameters and results are saved. Defaults to TRUE.}

\item{...}{Additional arguments passed to the underlying clustering functions.}
}
\value{
Returns a list of two sub-lists named \code{clusterres} and \code{params}. \code{clusterres} contains nested list with the comparison groups (e.g., \link{condition A}\emph{vs}\link{condition B}).
Each sub-list within \code{clusterres} contains two nested lists: \code{findPathCluster_opt_inputs}, which includes the \code{clustMethod} and \code{minCS} parameters,
and \code{findPathClusterres}, which contains cluster information and similarity values. \code{params} contains \code{max_sign_pathway} and \code{min_leading_edge_threshold} information.
}
\description{
This function serves as a wrapper for the \code{findPathCluster} function in the aPEAR package.
It processes the output from the \code{apear_data_prep} function, filters pathways based on the number
of leading edge genes, and selects pathways with the most significant p-values. It employs dynamic
clustering, primarily using the Markov clustering method, adjusting the minimum cluster size as needed
to optimize the number of labels displayed on the final graph. If Markov clustering fails, it
automatically attempts hierarchical clustering as a fallback.
}
\note{
Ensure that the \code{output_dir} is specified if \code{save_cluster_parameters_and_results} is set to TRUE.
The parameter \code{save_cluster_parameters_and_results} is strongly encouraged to be set as \code{TRUE} to ensure the reproducibility.
If the clusters for the final graph are too crowded, you may consider to decrease \code{max_sign_pathway}.
}
\examples{
# Assuming `apear_data_prep_output` is your prepared input list:
findPathClusterres <- apear_find_clusters(apear_input_list,
                                          max_sign_pathway = 500,
                                          min_leading_edge_threshold = 2,
                                          output_dir = "path/to/save",
                                          save_cluster_parameters_and_results=TRUE)

}
