% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPEAR_wrapper_functions.R
\name{apear_plot_cluster}
\alias{apear_plot_cluster}
\title{Plot aPEAR Clusters}
\usage{
apear_plot_cluster(
  apear_input_list,
  findPathClusterres,
  color_mapping = NULL,
  repelLabels = FALSE,
  drawEllipses = FALSE,
  fontSize = 2.5,
  show_parameters = TRUE,
  output_dir = NULL,
  output_width = 10,
  output_height = 10,
  save_network_plot = TRUE,
  save_apearres_obj = TRUE,
  ...
)
}
\arguments{
\item{apear_input_list}{The updated list output from the \code{apear_update_and_save_input} function.}

\item{findPathClusterres}{The list output from the \code{apear_find_clusters} function containing
clustering results.}

\item{color_mapping}{A named vector of colors to be assigned to comparison groups for plotting. \code{NULL} by default.}

\item{repelLabels}{Logical; if \code{TRUE}, labels are adjusted to minimize overlap in the plot. Defaults to \code{FALSE}.}

\item{drawEllipses}{Logical; if \code{TRUE}, ellipses are drawn around the clusters. Defaults to \code{FALSE}.}

\item{fontSize}{Numeric; the font size used for cluster labels in the plot. Defaults to 2.5.}

\item{show_parameters}{Logical; if \code{TRUE}, plot includes clustering parameters. Defaults to \code{TRUE}.}

\item{output_dir}{A string specifying the directory where output files will be saved. Defaults to \code{TRUE}.}

\item{output_width}{Numeric; the width of the output plot. Defaults to 10.}

\item{output_height}{Numeric; the height of the output plot. Defaults to 10.}

\item{save_network_plot}{Logical; if \code{TRUE}, the network plot is saved to the specified \code{output_dir}. Defaults to \code{TRUE}.}

\item{save_apearres_obj}{Logical; if \code{TRUE}, the aPEAR results object is saved to the specified \code{output_dir}. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to the underlying clustering functions.}
}
\value{
Generates and optionally saves a visual representation of the aPEAR clusters and the aPEAR
results object.
}
\description{
Visualizes clusters generated by aPEAR analysis using the updated aPEAR input list and the
clustering results. The function allows customization of the plot appearance and the option
to save the plot and aPEAR results object.
}
\examples{
# Assuming `apear_input_list` and `findPathClusterres` have been defined, and color mapping is set:
apear_network_list <-  apear_plot_cluster(apear_input_list_updated,
                                          findPathClusterres,
                                          color_mapping = c(Group1 = "blue", Group2 = "red"),
                                          repelLabels = FALSE,
                                          drawEllipses = FALSE,
                                          fontSize = 2.5,
                                          show_parameters = TRUE,
                                          output_dir = "path/to/save",
                                          output_width = 10,
                                          output_height = 10,
                                          save_network_plot = TRUE,
                                          save_apearres_obj = TRUE)

}
