% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_helperfunctions.R
\name{calculate_percent.hemoglobin}
\alias{calculate_percent.hemoglobin}
\title{Calculate hemoglobin features in Seurat object}
\usage{
calculate_percent.hemoglobin(sobj, hemoglobin.features)
}
\arguments{
\item{sobj}{Seurat object}

\item{hemoglobin.features}{character vector, hemoglobin gene symbols to search in dataset, default will search all mouse and human hemoglobin gene symbols}
}
\value{
data.frame with percent.hemoglobin in all cells
}
\description{
Calculate hemoglobin features in Seurat object
}
\examples{
sobj[['percent.hemoglobin']] <- Seurat::PercentageFeatureSet(sobj, features = hemoglobin.features)
}
