% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_config.R
\name{.annotate_sobjlist_for_integration}
\alias{.annotate_sobjlist_for_integration}
\title{Add Code, Sample, Condition, and Barcode columns before integration prep.}
\usage{
.annotate_sobjlist_for_integration(sobjlist, sample_metadata)
}
\description{
Add Code, Sample, Condition, and Barcode columns before integration prep.
}
\keyword{internal}
