% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_config.R
\name{.append_concat_rna_assay}
\alias{.append_concat_rna_assay}
\title{Attach joined raw RNA counts assay to integrated object for pseudobulk DE.}
\usage{
.append_concat_rna_assay(sobjint, outdir_int_objects, sample_metadata)
}
\description{
Attach joined raw RNA counts assay to integrated object for pseudobulk DE.
}
\keyword{internal}
