% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_config.R
\name{.build_matlist_from_sobjlist}
\alias{.build_matlist_from_sobjlist}
\title{Extract RNA counts per sample from sobjlist for RISC concat matrix (input_seurat_obj path).}
\usage{
.build_matlist_from_sobjlist(sobjlist, sample_metadata)
}
\description{
Extract RNA counts per sample from sobjlist for RISC concat matrix (input_seurat_obj path).
}
\keyword{internal}
