% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.compute_stability_metrics}
\alias{.compute_stability_metrics}
\title{Aggregate stability metrics across the parameter field}
\usage{
.compute_stability_metrics(
  globref_l,
  repres_flat,
  bsindex,
  paramfield,
  verbose = FALSE,
  stability_workernum = 1L
)
}
\description{
Aggregate stability metrics across the parameter field
}
\keyword{internal}
