% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr_cache.R
\name{.default_msigdbr_pwaycats}
\alias{.default_msigdbr_pwaycats}
\title{Default MSigDB pathway categories used by scDAPP}
\usage{
.default_msigdbr_pwaycats()
}
\value{
Character vector of \code{gs_subcat} values after normalization.
}
\description{
Default MSigDB pathway categories used by scDAPP
}
\keyword{internal}
