% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_config.R
\name{.finalize_sobjint_clusters}
\alias{.finalize_sobjint_clusters}
\title{Set active ident, default assay, and misc metadata after integration clustering.}
\usage{
.finalize_sobjint_clusters(sobjint, config, int_clust_lab)
}
\description{
Set active ident, default assay, and misc metadata after integration clustering.
}
\keyword{internal}
