% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_run.R
\name{.get_integrate_layers_fn}
\alias{.get_integrate_layers_fn}
\title{Map integration_method string to Seurat::IntegrateLayers function.}
\usage{
.get_integrate_layers_fn(method_name)
}
\description{
Map integration_method string to Seurat::IntegrateLayers function.
}
\keyword{internal}
