% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr_cache.R
\name{.msigdb_celltype_gs_subcat}
\alias{.msigdb_celltype_gs_subcat}
\title{MSigDB gs_subcat used for cluster-marker cell-type ORA (currently C8)}
\usage{
.msigdb_celltype_gs_subcat()
}
\value{
Single normalized \code{gs_subcat} string.
}
\description{
MSigDB gs_subcat used for cluster-marker cell-type ORA (currently C8)
}
\keyword{internal}
