% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.parse_stability_param_id}
\alias{.parse_stability_param_id}
\title{Parse PCs_1-N.res_R stability grid id into pcs_int and res_int integers.}
\usage{
.parse_stability_param_id(params_i)
}
\description{
Parse PCs_1-N.res_R stability grid id into pcs_int and res_int integers.
}
\keyword{internal}
