% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.pick_best_stability_params}
\alias{.pick_best_stability_params}
\title{Select best PC/res from stability summary table}
\usage{
.pick_best_stability_params(perparam_meanscores)
}
\description{
Select best PC/res from stability summary table
}
\keyword{internal}
