% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.plot_stability_bootstrap_ari}
\alias{.plot_stability_bootstrap_ari}
\title{Boxplot of bootstrap ARI distributions per params_i vs reference clustering.}
\usage{
.plot_stability_bootstrap_ari(perbootstrap, perparam)
}
\description{
Boxplot of bootstrap ARI distributions per params_i vs reference clustering.
}
\keyword{internal}
