% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.plot_stability_nclust_ref}
\alias{.plot_stability_nclust_ref}
\title{Bar chart of reference cluster counts per params_i (full-data integration).}
\usage{
.plot_stability_nclust_ref(perparam, selected_params_i)
}
\description{
Bar chart of reference cluster counts per params_i (full-data integration).
}
\keyword{internal}
