% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.prepare_stability_perparam}
\alias{.prepare_stability_perparam}
\title{Add combinedscore and factor order for stability summary plotting.}
\usage{
.prepare_stability_perparam(perparam)
}
\description{
Add combinedscore and factor order for stability summary plotting.
}
\keyword{internal}
