% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_config.R
\name{.read_sobjlist_from_dir}
\alias{.read_sobjlist_from_dir}
\title{Load Code.rds objects from pipeline tmpobjdir for Seurat merge integration.}
\usage{
.read_sobjlist_from_dir(objdir, codes)
}
\description{
Load Code.rds objects from pipeline tmpobjdir for Seurat merge integration.
}
\keyword{internal}
