% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_resume.R
\name{.resume_can_load}
\alias{.resume_can_load}
\title{Whether a stage can be loaded from cache.}
\usage{
.resume_can_load(
  stage,
  outdir,
  fingerprint,
  sample_codes = NULL,
  stability_outdir = NULL,
  outdir_int = NULL,
  engine = "RISC"
)
}
\description{
Whether a stage can be loaded from cache.
}
\keyword{internal}
