% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_resume.R
\name{.resume_restore_tmpobjdir}
\alias{.resume_restore_tmpobjdir}
\title{Copy resume sample RDS into tmpobjdir for update_indi_objects / Seurat int.}
\usage{
.resume_restore_tmpobjdir(outdir, tmpobjdir, sample_codes)
}
\description{
Copy resume sample RDS into tmpobjdir for update_indi_objects / Seurat int.
}
\keyword{internal}
