% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_resume.R
\name{.resume_sample_key}
\alias{.resume_sample_key}
\title{Sample metadata columns used in fingerprints.}
\usage{
.resume_sample_key(sample_metadata)
}
\description{
Sample metadata columns used in fingerprints.
}
\keyword{internal}
