% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_align_sample_metadata}
\alias{.risc_align_sample_metadata}
\title{Align \code{sample_metadata} rows to \code{names(risclist)}.}
\usage{
.risc_align_sample_metadata(risclist, sample_metadata)
}
\description{
Align \code{sample_metadata} rows to \code{names(risclist)}.
}
\keyword{internal}
