% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_lookup_reference_index}
\alias{.risc_lookup_reference_index}
\title{Index of a Code/Sample in sample_metadata.}
\usage{
.risc_lookup_reference_index(sample_metadata, risc_reference, required = TRUE)
}
\description{
Index of a Code/Sample in sample_metadata.
}
\keyword{internal}
