% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_put_reference_first}
\alias{.risc_put_reference_first}
\title{Put the reference RISC object first (RISC \code{scMultiIntegrate} convention).}
\usage{
.risc_put_reference_first(risclist, ref)
}
\description{
Put the reference RISC object first (RISC \code{scMultiIntegrate} convention).
}
\keyword{internal}
