% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_reference_fingerprint_value}
\alias{.risc_reference_fingerprint_value}
\title{Canonical \code{risc_reference} string for resume fingerprints.}
\usage{
.risc_reference_fingerprint_value(risc_reference = NULL)
}
\description{
Canonical \code{risc_reference} string for resume fingerprints.
}
\keyword{internal}
