% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_reference_selection_read}
\alias{.risc_reference_selection_read}
\title{Read RISC reference selection artifacts if present.}
\usage{
.risc_reference_selection_read(dir)
}
\description{
Read RISC reference selection artifacts if present.
}
\keyword{internal}
