% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_reference_selection_write}
\alias{.risc_reference_selection_write}
\title{Write RISC reference selection artifacts.}
\usage{
.risc_reference_selection_write(selection, dir)
}
\description{
Write RISC reference selection artifacts.
}
\keyword{internal}
