% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_reference_table_autov2}
\alias{.risc_reference_table_autov2}
\title{Build the autoV2 HTML/CSV table.}
\usage{
.risc_reference_table_autov2(scores, pick, selected_code = NULL)
}
\description{
Build the autoV2 HTML/CSV table.
}
\keyword{internal}
