% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.risc_reference_table_v1}
\alias{.risc_reference_table_v1}
\title{Build the legacy auto HTML/CSV table.}
\usage{
.risc_reference_table_v1(refscore, selected_code)
}
\description{
Build the legacy auto HTML/CSV table.
}
\keyword{internal}
