% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.risc_stability_cluster_at}
\alias{.risc_stability_cluster_at}
\title{Louvain cluster RISC integrated object at one stability grid point.}
\usage{
.risc_stability_cluster_at(
  risc_state,
  pcs_int,
  res,
  RISC_louvain_neighbors = 10L
)
}
\description{
Louvain cluster RISC integrated object at one stability grid point.
}
\keyword{internal}
