% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.risc_stability_integrate}
\alias{.risc_stability_integrate}
\title{Integrate RISC objects at given PC count}
\usage{
.risc_stability_integrate(
  risclist,
  sample_metadata,
  pcs_int,
  risc_reference = NULL,
  risc_ncore = 1L
)
}
\description{
Integrate RISC objects at given PC count
}
\keyword{internal}
