% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.risc_stability_integrate_cluster_at}
\alias{.risc_stability_integrate_cluster_at}
\title{Integrate RISC at pcs_int then cluster at res_int for stability reference run.}
\usage{
.risc_stability_integrate_cluster_at(
  risclist,
  sample_metadata,
  pcs_int,
  res_int,
  risc_reference = NULL,
  risc_ncore = 1L,
  RISC_louvain_neighbors = 10L
)
}
\description{
Integrate RISC at pcs_int then cluster at res_int for stability reference run.
}
\keyword{internal}
