% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_run.R
\name{.sctransform_merged_split_rna}
\alias{.sctransform_merged_split_rna}
\title{SCTransform merged object with split RNA layers for SCT IntegrateLayers path.}
\usage{
.sctransform_merged_split_rna(obj)
}
\description{
SCTransform merged object with split RNA layers for SCT IntegrateLayers path.
}
\keyword{internal}
