% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_run.R
\name{.seurat_cluster_umap_on_reduction}
\alias{.seurat_cluster_umap_on_reduction}
\title{Louvain cluster and UMAP on batch-corrected reduction after IntegrateLayers.}
\usage{
.seurat_cluster_umap_on_reduction(
  obj,
  config,
  npcs,
  res_int,
  cluster_name = NULL
)
}
\description{
Louvain cluster and UMAP on batch-corrected reduction after IntegrateLayers.
}
\keyword{internal}
