% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_run.R
\name{.seurat_integrate_layers_at_pcs}
\alias{.seurat_integrate_layers_at_pcs}
\title{Run normalization, PCA, and IntegrateLayers at npcs without clustering.}
\usage{
.seurat_integrate_layers_at_pcs(obj, config, npcs)
}
\description{
Run normalization, PCA, and IntegrateLayers at npcs without clustering.
}
\keyword{internal}
