% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_run.R
\name{.seurat_merge_from_tmpobjdir}
\alias{.seurat_merge_from_tmpobjdir}
\title{Merge per-sample RDS from tmpobjdir into one object for Seurat integration.}
\usage{
.seurat_merge_from_tmpobjdir(tmpobjdir, sample_metadata)
}
\description{
Merge per-sample RDS from tmpobjdir into one object for Seurat integration.
}
\keyword{internal}
