% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_backend_for_config}
\alias{.stability_backend_for_config}
\title{Seurat object backend from merged tmp objects}
\usage{
.stability_backend_for_config(
  config,
  sample_metadata,
  tmpobjdir = NULL,
  mdlist = NULL,
  datadir = NULL,
  matlist_path = NULL,
  risc_reference = NULL,
  risc_ncore = 1L,
  input_seurat_obj = FALSE,
  RISC_louvain_neighbors = 10L
)
}
\description{
Seurat object backend from merged tmp objects
}
\keyword{internal}
