% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_bootstrap_worker_seurat}
\alias{.stability_bootstrap_worker_seurat}
\title{Bootstrap replicate worker (Seurat): integrate per PC, cluster all res at that PC}
\usage{
.stability_bootstrap_worker_seurat(
  repi,
  paramfield,
  path_sobj_merged_temp,
  sweep_config,
  samplecolumn,
  propcells.perrep,
  bootstrapresdir,
  downsize_seurat,
  verbose = FALSE
)
}
\description{
Bootstrap replicate worker (Seurat): integrate per PC, cluster all res at that PC
}
\keyword{internal}
