% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_effective_workers}
\alias{.stability_effective_workers}
\title{Cap foreach worker count by number of stability tasks to run.}
\usage{
.stability_effective_workers(n_tasks, stability_workernum)
}
\description{
Cap foreach worker count by number of stability tasks to run.
}
\keyword{internal}
