% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_extract_cluster_df}
\alias{.stability_extract_cluster_df}
\title{Extract barcode/cluster data.frame from stability cluster column for metrics.}
\usage{
.stability_extract_cluster_df(meta, params_i)
}
\description{
Extract barcode/cluster data.frame from stability cluster column for metrics.
}
\keyword{internal}
