% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_foreach_packages}
\alias{.stability_foreach_packages}
\title{Foreach .packages vector; omit scDAPP when dev pkg root is set (already loaded)}
\usage{
.stability_foreach_packages(extra = "scDAPP")
}
\description{
Foreach .packages vector; omit scDAPP when dev pkg root is set (already loaded)
}
\keyword{internal}
