% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_global_ref_worker_risc}
\alias{.stability_global_ref_worker_risc}
\title{Global reference: integrate+cluster for one PC/res (RISC; worker loads risclist)}
\usage{
.stability_global_ref_worker_risc(
  params_i,
  path_risclist_temp,
  sample_metadata,
  risc_reference,
  risc_ncore,
  risc_louvain_neighbors,
  globrefdir,
  verbose = FALSE
)
}
\description{
Global reference: integrate+cluster for one PC/res (RISC; worker loads risclist)
}
\keyword{internal}
