% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_global_ref_worker_seurat}
\alias{.stability_global_ref_worker_seurat}
\title{Global reference: integrate+cluster for one PC/res (Seurat; merged counts)}
\usage{
.stability_global_ref_worker_seurat(
  params_i,
  path_sobj_merged_temp,
  sweep_config,
  globrefdir,
  verbose = FALSE
)
}
\description{
Global reference: integrate+cluster for one PC/res (Seurat; merged counts)
}
\keyword{internal}
