% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_risc_backend}
\alias{.stability_risc_backend}
\title{RISC stability backend closures}
\usage{
.stability_risc_backend(
  config,
  mdlist,
  sample_metadata,
  tmpobjdir = NULL,
  matlist_path = NULL,
  datadir = NULL,
  risc_reference = NULL,
  risc_ncore = 1L,
  input_seurat_obj = FALSE,
  RISC_louvain_neighbors = 10L
)
}
\description{
RISC stability backend closures
}
\keyword{internal}
