% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risc_reference.R
\name{.stability_risc_freeze_reference}
\alias{.stability_risc_freeze_reference}
\title{Freeze RISC reference on a stability backend after \code{prep_reference()}.}
\usage{
.stability_risc_freeze_reference(backend, outdir, verbose = TRUE)
}
\description{
Freeze RISC reference on a stability backend after \code{prep_reference()}.
}
\keyword{internal}
