% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_sweep_cluster_name}
\alias{.stability_sweep_cluster_name}
\title{Temporary cluster column name during stability sweep for one grid point.}
\usage{
.stability_sweep_cluster_name(params_i)
}
\description{
Temporary cluster column name during stability sweep for one grid point.
}
\keyword{internal}
