% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_stability.R
\name{.stability_sweep_save_params}
\alias{.stability_sweep_save_params}
\title{Save integrate+cluster results grouped by PC (integrate once per PC, cluster all res)}
\usage{
.stability_sweep_save_params(
  backend,
  params_to_run,
  savedir,
  filename_prefix = "",
  verbose = FALSE
)
}
\description{
Save integrate+cluster results grouped by PC (integrate once per PC, cluster all res)
}
\keyword{internal}
