% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pipeline_modules.R
\name{normalize_comps}
\alias{normalize_comps}
\title{Normalize comps for cross-condition DE and compositional modules}
\usage{
normalize_comps(comps)
}
\arguments{
\item{comps}{data.frame with c0, c1, and optional formula, contrast, label}
}
\value{
normalized comps data.frame
}
\description{
Normalize comps for cross-condition DE and compositional modules
}
