% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pipeline_modules.R
\name{ora_results_by_cluster}
\alias{ora_results_by_cluster}
\title{Filter flat ORA results}
\usage{
ora_results_by_cluster(
  ora_results,
  label,
  cluster = NULL,
  pathway_category = NULL
)
}
\arguments{
\item{ora_results}{output of \code{ORA_crosscondition_module()}}

\item{label}{comparison label}

\item{cluster}{optional cluster filter}

\item{pathway_category}{optional MSigDB category filter}
}
\value{
filtered data.frame
}
\description{
Filter flat ORA results
}
