% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pipeline_modules.R
\name{pathway_results_by_cluster}
\alias{pathway_results_by_cluster}
\title{Filter flat pathway GSEA results}
\usage{
pathway_results_by_cluster(
  pathway_results,
  label,
  cluster = NULL,
  pathway_category = NULL
)
}
\arguments{
\item{pathway_results}{output of \code{pathwayanalysis_crosscondition_module()}}

\item{label}{comparison label}

\item{cluster}{optional cluster filter}

\item{pathway_category}{optional MSigDB category filter}
}
\value{
filtered data.frame
}
\description{
Filter flat pathway GSEA results
}
