% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_RISC_scCluster_louvain_res.R
\name{scCluster_louvain_res}
\alias{scCluster_louvain_res}
\title{Cluster RISC object with Louvain Resolution option}
\usage{
scCluster_louvain_res(
  object,
  slot = "cell.umap",
  neighbor = 10,
  resolution = 1,
  algorithm = "kd_tree",
  npc = 20,
  redo = TRUE,
  random.seed = 123
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{slot}{The dimension_reduction slot for cell clustering. The default is "cell.umap" under RISC object "DimReduction" item for UMAP method, but the customer can add new dimension_reduction method under DimReduction and use it.}

\item{neighbor}{The neighbor cells for "igraph" method.}

\item{resolution}{Optional resolution parameter that allows the user to adjust the resolution parameter of the modularity function that the algorithm uses internally. Lower values typically yield fewer, larger clusters. The original definition of modularity is recovered when the resolution parameter is set to 1.}

\item{algorithm}{The algorithm for knn, the default is "kd_tree", all options: "kd_tree", "cover_tree", "CR", "brute".}

\item{npc}{The number of PCA or PLS used for cell clustering.}

\item{redo}{Whether re-cluster the cells.}

\item{random.seed}{The random seed, the default is 123}
}
\value{
RISC single cell dataset, the cluster slot; the column in coldata will be named as "Cluster_res.(resolution)"
}
\description{
This function is almost identical to \code{RISC::scCluster()} function with method = "louvain", with the addition of louvain resolution as a parameter. Higher resolution = more clusters. The R package \code{igraph} from CRAN must be installed and up to date, tested with version 1.4.3. Additionally, clusters will now be sorted biggest to smallest by number of cells (1 is biggest). The Cluster variable added to the object coldata will include resolution in the column name. Density clustering based on low dimensional embedding (ie UMAP) is not allowed in this function, only Louvain is supported, just use the original RISC verison for that, though it is not recommended.
}
\examples{
\dontrun{
obj0 = scCluster(obj0, slot = 'cell.pls', resolution = 0.5)
DimPlot(obj0, slot = "cell.umap", colFactor = 'Cluster_res.0.5', size = 2)
}
}
