% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_helperfunctions.R
\name{theme_dimplot}
\alias{theme_dimplot}
\title{a minimalistic theme for publication-quality nonlinear dimensional reduction plots like tSNE/UMAP}
\usage{
theme_dimplot(titlesize = 15)
}
\arguments{
\item{titlesize}{numeric, size of font}
}
\value{
A \code{ggplot2} theme object.
}
\description{
numeric axes are supposed to be not really meaningful in tSNE/UMAP
}
\examples{
Seurat::DimPlot(object = seurat_object) + theme_dimplot()
}
