% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pipeline_modules.R
\name{compositional_analysis_module}
\alias{compositional_analysis_module}
\title{Compositional analysis comparing proportional abundnace across conditions for integrated Seurat objects}
\usage{
compositional_analysis_module(
  sobjint,
  comps,
  sample_metadata,
  outdir_int,
  grouping_variable,
  compositional_test,
  fill_barplots
)
}
\arguments{
\item{sobjint}{integrated Seurat object. Metadata should have two columns: "Condition" corresponding to the A vs B conditions to compare across, and "Code" corresponding to sample / replicate names. A third column for clusters or celltypes should also be in the metadata and the name of that column will be passed to the \code{grouping_variable} parameter.}

\item{comps}{data.frame with two columns called c1, c2; these are the conditions you want to set up. Should be present in psuedobulk_metadata Condition column.}

\item{sample_metadata}{data.frame with three columns called Sample, Condition, Code.}

\item{outdir_int}{string, path to save results to. Will create a sub-directory called "compositional_proportion_analysis" and save inside of there.}

\item{grouping_variable}{string, column name of identity in Seurat object meta.data to stratify DE by. For example, clusters or celltype. Will perform A vs B DE in each of these groupings. Default is "seurat_clusters"}

\item{compositional_test}{string, which test to use, either "propeller" or "chisq"; will use chisq if not set and issue a warning}

\item{fill_barplots}{T/F, whether to "fill" the annotation barplots on the side of the heatmap, normalizing to 1; default = T}
}
\value{
will return a list with two items: the A vs B compositional analysis in the "composition_comps" element, with a sub-element for each comparison, and some global compositional information in the "globalcomposition" element.
}
\description{
This is a modular component of the scDAPP scRNAseq pipeline. Perform compositional analysis across conditions to compare the proportion of cell types. Supports multiple condtions (A vs B vs C). Supports "pseudobulk" replicate-aware analysis as implemented in the propeller test with arcsin transformation, as recommended by Simmons 2022 (https://doi.org/10.1101/2022.02.04.479123). Alternatively supports old-school, non-replicate aware chisq test as implemented by the \code{prop.test()} function.
}
\examples{
\dontrun{


# `sample_metadata` looks like this:
Sample,Condition,Code
SampleXYZ1,Control,Control1
SampleXYZ2,Control,Control2
SampleABC1,KO1,KO1_1
SampleABC2,KO1,KO1_2
SampleJKL1,KO2,KO2_1
SampleJKL2,KO2,KO2_1


# `comps` looks like this:
c1,c2
KO1,Control
KO2,Control
KO1,K2


## Run the analysis ##
comp_result <- compositional_analysis_module(sobjint,
comps,
sample_metadata,
outdir_int,
grouping_variable,
compositional_test = 'propeller')

#Get the output for KO1 vs Control; replace KO1_vs_Control to get the other comparisons
# differential composition results - heatmap
comp_result$composition_comps$KO1_vs_Control$hmprop_comp

# differential composition results - table
comp_result$composition_comps$KO1_vs_Control$compres

#Get some global information including cell numbers, proportions
# table of cell numbers
comp_result$globalcomposition$cellstab

#table of cell proportions, ie the table above with each column divided by column sum
comp_result$globalcomposition$proptab

#heatmap of cell proportions, without any statistical testing
comp_result$globalcomposition$hmprop
}

}
