% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pipeline_modules.R
\name{ORA_crosscondition_module}
\alias{ORA_crosscondition_module}
\title{Overrepresentation analysis for cross condition pathway analysis}
\usage{
ORA_crosscondition_module(
  m_bycluster_crosscondition_de_comps,
  pathways,
  sample_metadata,
  comps,
  crossconditionDE_padj_thres,
  crossconditionDE_lfc_thres,
  crossconditionDE_min.pct,
  pathway_padj_thres,
  pwaycats,
  workernum,
  outdir_int
)
}
\arguments{
\item{m_bycluster_crosscondition_de_comps}{the output of \code{scDAPP::de_across_conditions_module()}.}

\item{pathways}{data.frame, the output of \code{scDAPP::preppathways_pathwayanalysis_crosscondition_module()}}

\item{sample_metadata}{data.frame with sample names and conditions, same as in \code{scDAPP::de_across_conditions_module()}, see that function's documentation for description.}

\item{comps}{data.frame with conditions to test in GSEA, same as in \code{scDAPP::de_across_conditions_module()}, see that function for description}

\item{pathway_padj_thres}{numeric, threshold for significance of pathway enrichment after multiple test correction, passed to qvalueCutoff in \code{clusterProfiler::enricher}}

\item{pwaycats}{UNTESTED CURRENTLY. character vector of msigdb pathways data.frame in gs_subcat to run.}

\item{workernum}{integer. number of CPUs. default = 1.}

\item{outdir_int}{string, directory to save pathways to. Will create a sub-directory called "overrepresentation_pathway_analysis" and save inside of there.}
}
\value{
a list that contains the raw results and plots as a nested loop: first level is A vs B comparison, then category-by-category of MSIGDB database, then a data.frame of cluster-by-cluster results
}
\description{
This function is a modular component of the scRNAseq pipeline. Perform OverRepresentation Analysis (ORA) via the ClusterProfiler package on the results of differential expression (DE) analysis for cross-condition comparison. Multiple conditions are supported. ClusterProfiler objects and tables are saved.
}
\examples{
\dontrun{


# FIRST: run `scDAPP::de_across_conditions_module()`. the output object of that is used as the main input for this pathway analysis function.

# SECOND: prep pathways before running
pathways <- scDAPP::preppathways_pathwayanalysis_crosscondition_module(species = species,
outdir_int = outdir_int)

# THIRD: run the pathway analysis.
# see `scDAPP::de_across_conditions_module()` for a description of the sample_metadata and comps files.
pways_output_list <- scDAPP::ORA_crosscondition_module(
m_bycluster_crosscondition_de_comps = m_bycluster_crosscondition_de_comps,
pathways = pathways,
sample_metadata = sample_metadata,
comps = comps,
workernum = 6,
outdir_int = outdir_int
)

# Access the output files
# note that all results will be saved to outdir_int.
pathway_analysis_mainlist_comps <- pways_output_list$pathway_analysis_mainlist_comps
pathwaysummplots_comps <- pways_output_list$pathwaysummplots_comps

#these are both lists that contain the raw tables and per-cluster plots (first object);
# and summaryplots that show pathway enrichment across clusters (second object).

}
}
