% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_helperfunctions.R
\name{pdftable}
\alias{pdftable}
\title{Make a table that can print to a pdf page}
\usage{
pdftable(tabledf, title, titlesize, padding)
}
\arguments{
\item{tabledf}{data.frame to put as table on pdf}

\item{title}{title for table}

\item{titlesize}{title font size, default is 15}

\item{padding}{whitespace between title and table, default=1}
}
\description{
Make a table that can print to a pdf page
}
