% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pipeline_modules.R
\name{preppathways_pathwayanalysis_crosscondition_module}
\alias{preppathways_pathwayanalysis_crosscondition_module}
\title{Prep MSIGDB pathways for pathway analysis}
\usage{
preppathways_pathwayanalysis_crosscondition_module(species, outdir_int)
}
\arguments{
\item{species}{string, species such as "Homo sapeins" or "Mus musculus"}

\item{outdir_int}{string, directory to save pathways to. Will create a sub-directory called "pathwayanalysis_crosscondition" and save inside of there. We save pathways since the database updates over time.}
}
\value{
a data.frame similar to the output of \code{msigdbr::msigdbr}, but filtering for some specific categories / subcategories.
}
\description{
This is a modular component of the scRNAseq analysis pipeline. Prep pathways from MSIGDB via the msigdbr package. We include the Hallmarks category; Gene Ontology BP, MF and CC; Reactome; KEGG; transcription factor CHIP-seq targets in the Gene Transcription Regulation Database (TFT_GTRD); inferred transcription factor targets via motif analysis from Xie et al Nature 2005 (TFT_Legacy). Only gene sets with < 500 genes are included.
}
\examples{
\dontrun{
pathways <- preppathways_pathwayanalysis_crosscondition_module(
species = 'Mus musculus',
outdir_int = 'path/to/directory')
}
}
