% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_helperfunctions.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulk Seurat objects at whole sample or celltype level}
\usage{
pseudobulk(
  obj,
  grouping_colname_in_md,
  metadata,
  rawh5_path,
  assay,
  slot,
  min_cells
)
}
\arguments{
\item{obj}{\itemize{
\item seurat object, or a matrix
}}

\item{grouping_colname_in_md}{\itemize{
\item optional, a string, the column name of \code{sobj@meta.data} (if obj is a seurat object) or metadata (if using matrix and metadata input) to use as "cell types" or any other grouping for pseudobulking at group level. if not provided, pseudobulk the entire matrix. default, not used.
}}

\item{metadata}{\itemize{
\item data.frame with cell metadata, similar to \code{seuratobject@meta.data}. pass this only if
}}

\item{rawh5_path}{\itemize{
\item optional, a string, the path to a rawH5 file, if provided will use the raw matrix subsetted by cells in sobj; if not will just pseudobulk the seurat object. useful if some filtering was applied to seurat object but you want to pseudobulk the whole matrix without that filtering, but with only using cells in seurat object. Default is not to use this.
}}

\item{assay}{\itemize{
\item optional, a string, the name of the Seurat object assay to pull matrix from if rawh5_path is not provided. Default is "RNA" assay
}}

\item{slot}{\itemize{
\item optional, a string, the name of the Seurat object slot within the designated object assay to pull matrix from if rawh5_path is not provided. Default is "counts" slot
}}

\item{min_cells}{\itemize{
\item numeric, minimum number of cells in the cluster to pseudobulk. default is 7 cells.
}}
}
\value{
a data.frame. if grouping_colname_in_md is provided, each celltype will have a pseudobulked column, if not the data.frame will just be one column for the whole sample matrix.
}
\description{
Pseudobulking is performed adding up gene expression values for each cell, for either cell types or whole sample.
}
