% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_runner.R
\name{r_package_test}
\alias{r_package_test}
\title{Test installation of the packages}
\usage{
r_package_test()
}
\value{
a data.frame with packages and version numbers
}
\description{
This will load the key dependencies and report the versions of those dependencies.
}
\examples{
\dontrun{
SpatialSingulomics::r_package_test()
}
}
